develop -> small fixes

+++ src/main/java/com/kiwi/observability/MetricsRegistry.java
@@ -8,7 +8,7 @@ import java.util.concurrent.ConcurrentMap;
 public final class MetricsRegistry {
     private static final MetricsRegistry instance = new MetricsRegistry();
 
-    private final ConcurrentMap<String, Integer> gauges = new ConcurrentHashMap<>();
+    private final ConcurrentMap<String, Long> gauges = new ConcurrentHashMap<>();
     private final ConcurrentMap<String, Long> counters = new ConcurrentHashMap<>();
 
     private final long startUpMillis;
@@ -18,15 +18,15 @@ public final class MetricsRegistry {
     }
 
     public void registerGauge(String name) {
-        gauges.putIfAbsent(name, 0);
+        gauges.putIfAbsent(name, 0L);
     }
 
     public void registerCounter(String name) {
         counters.putIfAbsent(name, 0L);
     }
 
-    public void updateGauge(String name, int delta) {
-        gauges.merge(name, delta, Integer::sum);
+    public void updateGauge(String name, long delta) {
+        gauges.merge(name, delta, Long::sum);
     }
 
     public void updateCounter(String name) {
@@ -37,7 +37,7 @@ public final class MetricsRegistry {
         counters.merge(name, delta, Long::sum);
     }
 
-    public Map<String, Integer> getGauges() {
+    public Map<String, Long> getGauges() {
         return Collections.unmodifiableMap(gauges);
     }
