+++ src/main/java/com/kiwi/observability/dto/MetricsDataDto.java
@@ -8,7 +8,7 @@ import java.util.Map;
 public record MetricsDataDto(
     String protocolVersion,
     String infoSchemaVersion,
-    Map<String, Integer> gauges,
+    Map<String, Long> gauges,
     Map<String, Long> counters,
     long serverStart,
     long serverUptime,
